+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug fixes and improvements
 (Add entries below in the order of server, console, cli, docs, others)
+- console: add foreign key CRUD functionality to ms sql server tables
 
 
 ## v2.0.0-beta.1